Refactoring. Move bundle status check to plugin. Move configurable plugin to around. Both facts comes from fact for bundle and configurable produts, salability status is depends only on salability status of their children (and does not depend on the condition chain).
